+++ GenericEventRunner.DomainParts/EntityEventsBase.cs
@@ -9,13 +9,16 @@ namespace GenericEventRunner.DomainParts
     /// <summary>
     /// This is a class that the EF Core entity classes inherit to add events
     /// </summary>
-    public abstract class EntityEventsBase : IEntityWithBeforeSaveEvents, IEntityWithAfterSaveEvents
+    public abstract class EntityEventsBase : IEntityWithBeforeSaveEvents, IEntityWithDuringSaveEvents, IEntityWithAfterSaveEvents
     {
         //Events are NOT stored in the database - they are transitory events
         //Events are created within a single DBContext and are cleared every time SaveChanges/SaveChangesAsync is called
         
         //This holds events that are run before SaveChanges is called
         private readonly List<IDomainEvent> _beforeSaveEvents = new List<IDomainEvent>();
+
+        //This holds events that are run within a transaction containing a call to SaveChanges 
+        private readonly List<IDomainEvent> _duringSaveEvents = new List<IDomainEvent>();
 
         //This holds events that are run after SaveChanges finishes successfully
         private readonly List<IDomainEvent> _afterSaveChangesEvents  = new List<IDomainEvent>();
@@ -24,9 +27,11 @@ namespace GenericEventRunner.DomainParts
         /// This allows an entity to add an event to this class
         /// </summary>
         /// <param name="dEvent">This is the domain event you want to sent</param>
-        /// <param name="eventToSend">This allows you to send the event to either BeforeSave list, the AfterSave list or both. Default is BeforeSave List</param>
+        /// <param name="eventToSend">This allows you to send the event to either BeforeSave, DuringSave or AfterSave. Default is BeforeSave List</param>
         public void AddEvent(IDomainEvent dEvent, EventToSend eventToSend = EventToSend.BeforeSave)
         {
+            if (eventToSend == EventToSend.DuringSave)
+                _duringSaveEvents.Add(dEvent);
             if (eventToSend == EventToSend.BeforeSave || eventToSend == EventToSend.BeforeAndAfterSave)
                 _beforeSaveEvents.Add(dEvent);
             if (eventToSend == EventToSend.AfterSave || eventToSend == EventToSend.BeforeAndAfterSave)
@@ -36,7 +41,6 @@ namespace GenericEventRunner.DomainParts
         /// <summary>
         /// This gets all the events in the BeforeSaveEvents list, and clears that list at the same time
         /// </summary>
-        /// <returns></returns>
         public ICollection<IDomainEvent> GetBeforeSaveEventsThenClear()
         {
             var eventCopy = _beforeSaveEvents.ToList();
@@ -45,14 +49,30 @@ namespace GenericEventRunner.DomainParts
         }
 
         /// <summary>
+        /// This returns the events that should be run within a transaction containing a call to SaveChanges
+        /// </summary>
+        public ICollection<IDomainEvent> GetDuringSaveEvents()
+        {
+            return _duringSaveEvents;
+        }
+
+        /// <summary>
+        /// This clears all the during save events once the code within the transaction has finished
+        /// </summary>
+        public void ClearDuringSaveEvents()
+        {
+            _duringSaveEvents.Clear();
+        }
+
+        /// <summary>
         /// This gets all the events in the AfterSaveEvents list, and clears that list at the same time
         /// </summary>
-        /// <returns></returns>
         public ICollection<IDomainEvent> GetAfterSaveEventsThenClear()
         {
             var eventCopy = _afterSaveChangesEvents.ToList();
             _afterSaveChangesEvents.Clear();
             return eventCopy;
         }
+
     }
 }