+++ GenericEventRunner.DomainParts/IEntityWithBeforeSaveEvents.cs
@@ -5,8 +5,14 @@ using System.Collections.Generic;
 
 namespace GenericEventRunner.DomainParts
 {
+    /// <summary>
+    /// Add this interface to an entity class to support BeforeSaveEvents
+    /// </summary>
     public interface IEntityWithBeforeSaveEvents
     {
+        /// <summary>
+        /// This gets all the events in the BeforeSaveEvents list, and clears that list at the same time
+        /// </summary>
         ICollection<IDomainEvent> GetBeforeSaveEventsThenClear();
     }
 }