+++ ReleaseNotes.md
@@ -1,5 +1,13 @@
 # Release notes
 
+## 2.1.0
+
+- BREAKING CHANGE - the definitions of the events handler has changed. The callingEntity parameter is now an object. 
+- New Feature: Added during event and event handlers. The event is called within a transaction.
+- New feature: DuringSaveStage attribute. Allows you move the during event from running after SaveChanges in the transaction to before the call to SaveChanges (but after DetectChanges has been called).
+- New Feature: You can now pick what types of events an entity supports by applying the approriate interfaces: IEntityWithBeforeSaveEvents, IEntityWithDuringSaveEvents, and IEntityWithAfterSaveEvents
+
+
 ## 2.0.0
 
 - Make GenericEventRunner support a clean archtecture